# Doc/openapi.yaml
openapi: 3.0.3
info:
  title: StockWise (FinApp26) Market Data API
  version: 1.0.0
  description: |
    Canonical read API for every screen in StockWise. This documents the **data
    contract** the frontend consumes — today implemented as direct Firestore
    reads (via `useCollection()` in `app/iq/hooks/useCollection.ts`) rather than
    a deployed REST gateway, so every path below maps 1:1 to a Firestore
    collection written by a NestJS sync job in `backend/src/sync/`. The shapes
    are copied verbatim from the actual job code (not from the earlier planning
    doc `Doc/05_API_Vendor_Tracker.md`, which has drifted from what is really
    implemented — e.g. `analyst_actions` was planned as a per-event
    upgrade/downgrade feed but is actually a consensus-vote snapshot today).

    **Status legend** (see `x-status` on every path):
    - `live` — job exists, writes real data, screen(s) already read it.
    - `planned` — no backend job exists yet; schema is a proposed shape based
      on the named vendor's typical payload. Nothing reads/writes this today.

    **Source legend** (see `x-primary-source` / `x-alternate-source` on every
    path): the vendor actually wired in `backend/src/vendors/`, plus a faster
    or more complete alternate and why it's faster — see each path's
    description for specifics.

    **Swapping a source never changes this document — and fallback is
    automatic and never silent.** `Company` and `MarketMover` (the two
...
    predates them.

...
servers:
...
paths:
...
  /sync/jobs:
    get:
      tags: [Ops]
      summary: List every registered sync job with running + last-run state
      x-status: live
      x-consumed-by:
        - "Not a StockWise screen — consumed by backendUI (backendUI/index.html), the separate static ops dashboard, not the Next.js app."
      operationId: listSyncJobs
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SyncJobStatus" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /sync/status:
    get:
      tags: [Ops]
      summary: Last-run status for every job (no running-state)
      x-status: live
      x-consumed-by:
        - "Not a StockWise screen — ops-only, same as GET /sync/jobs."
      operationId: getAllSyncStatus
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SyncMeta" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /sync/{job}/status:
    get:
      tags: [Ops]
      summary: Last-run status for one job
      x-status: live
      x-consumed-by:
        - "Not a StockWise screen — ops-only, same as GET /sync/jobs."
      operationId: getSyncStatus
      parameters:
        - $ref: "#/components/parameters/JobName"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SyncMeta" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /sync/{job}/run:
    post:
      tags: [Ops]
      summary: Manually trigger a sync job (for testing)
      x-status: live
      x-consumed-by:
        - "Not a StockWise screen — ops-only, same as GET /sync/jobs."
      operationId: runSyncJob
      parameters:
        - $ref: "#/components/parameters/JobName"
      responses:
        "200":
          description: Job ran; per-job result shape (counts vary by job)
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "404":
          description: Unknown job name
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "502": { $ref: "#/components/responses/UpstreamError" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
...
components:
  parameters:
...
    JobName:
      name: job
      in: path
      required: true
      schema:
        type: string
        enum:
          - companies
          - market-movers
          - sectors
          - market-indices
          - earnings
          - analyst-actions
          - news
          - sec-13f
          - sec-form4
          - ticker-universe
...
  responses:
...
    UpstreamError:
      description: >-
        The upstream vendor returned an error, malformed data, or timed out.
        If a fallback source is configured for this path (see
        `x-fallback-behavior`) it was already tried and also failed —
        `fallbackAttempted` lists every source tried and why each failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example:
            statusCode: 502
            error: UPSTREAM_VENDOR_ERROR
            message: "All sources failed for company NVDA: fmp (HTTP 500 from FMP profile endpoint, retryable), polygon (no data returned (ticker not found on this source), not retryable)"
            retryable: true
            fallbackAttempted:
              - source: fmp
                error: "HTTP 500 from FMP profile endpoint"
              - source: polygon
                error: "no data returned (ticker not found on this source)"
            timestamp: "2026-07-05T14:00:00Z"
    ServiceUnavailable:
      description: The backend itself (or Firestore) is unreachable — distinct from an upstream vendor failure.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example: { statusCode: 503, error: SERVICE_UNAVAILABLE, message: "Firestore write failed after 3 retries — see backend logs for the underlying gRPC error.", retryable: true, timestamp: "2026-07-05T14:00:00Z" }
...
  schemas:
...
    ErrorResponse:
      type: object
      description: >-
        The single error shape used by every 4xx/5xx response in this API.
        `message` is always populated with a complete, specific sentence —
        never empty, never a generic "An error occurred."
      required: [statusCode, error, message, timestamp]
      properties:
        statusCode: { type: integer }
        error:
          type: string
          description: Stable machine-readable code — safe to switch on in client code.
          enum:
            - INVALID_PARAMETER
            - UNAUTHENTICATED
            - PERMISSION_DENIED
            - NOT_FOUND
            - VENDOR_RATE_LIMITED
            - UPSTREAM_VENDOR_ERROR
            - SERVICE_UNAVAILABLE
            - NOT_IMPLEMENTED
        message: { type: string, description: "Human-readable, specific explanation — always non-empty." }
        source: { type: string, nullable: true, description: "Vendor/adapter that failed, when this error originated from a vendor call." }
        retryable: { type: boolean, nullable: true, description: "True if retrying later is likely to succeed (rate limit, timeout, 5xx). False for permanent conditions (not found, bad request)." }
        fallbackAttempted:
          type: array
          nullable: true
          description: Present only on UPSTREAM_VENDOR_ERROR when a fallback source was configured and also failed — every attempt, in order.
          items:
            type: object
            properties:
              source: { type: string }
              error: { type: string }
        timestamp: { type: string, format: date-time }
    SyncJobStatus:
      type: object
      description: >-
        GET /sync/jobs's per-job entry. `collections`/`cronExpression`/
        `timeZone` (added 2026-07-09) come from the in-process SyncRegistry
        (declared once per job at registration time, next to its
        @Cron(...)) — always populated even when Firestore itself is
        unreachable (see `metaError`), unlike lastSyncedAt/lastStatus/
        lastCount/lastError which depend on reading sync_meta.
      properties:
        name: { type: string, example: market-movers }
        isRunning: { type: boolean }
        runningSince: { type: string, format: date-time, nullable: true }
        lastSyncedAt: { type: string, format: date-time, nullable: true }
        lastStatus: { type: string, enum: [ok, error], nullable: true }
        lastCount: { type: integer, nullable: true }
        lastError: { type: string, nullable: true }
        collections: { type: array, items: { type: string }, example: ["market_movers", "market_movers_history"] }
        cronExpression: { type: string, example: "0 18 * * 1-5" }
        timeZone: { type: string, example: "America/New_York" }
        nextRunAt: { type: string, format: date-time, nullable: true, description: "Added 2026-07-09. Computed fresh on every call via the `cron` package's CronJob.nextDate() (the same library @nestjs/schedule's @Cron(...) uses internally, so this always agrees with when the job will actually fire) — never cached/stored, never re-derived by hand-rolled cron math. Null only if cronExpression itself fails to parse." }
        metaError: { type: string, nullable: true, description: "Set instead of the last* fields when sync_meta itself couldn't be read (e.g. Firestore auth failure) — the job still appears with its registry-known collections/schedule." }
    SyncMeta:
      type: object
      description: >-
        Firestore `sync_meta/{jobName}` — persisted on every run by
        SyncMetaService.record(). collections/cronExpression/timeZone
        (added 2026-07-09) are looked up from SyncRegistry at record time
        and written alongside the run result, so this doc alone answers
        "which table did this job affect, and how often does it run"
        without reading backend source.
      properties:
        job: { type: string }
        lastSyncedAt: { type: string, format: date-time, nullable: true }
        lastStatus: { type: string, nullable: true }
        lastCount: { type: integer, nullable: true }
        lastError: { type: string, nullable: true }
        collections: { type: array, items: { type: string }, nullable: true }
        cronExpression: { type: string, nullable: true }
        timeZone: { type: string, nullable: true }
...